bookbazar/src/app/auth/register/register.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ToastrService } from 'ngx-toastr';
import { AuthService } from 'src/app/shared/auth.service';
import { TestModule } from 'src/app/test/test.module';

import { RegisterComponent } from './register.component';

describe('RegisterComponent', () => {
  let component: RegisterComponent;
  let fixture: ComponentFixture<RegisterComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ RegisterComponent ],
      imports : [TestModule],
      providers : [AuthService,ToastrService]
    })
    .compileComponents();

    fixture = TestBed.createComponent(RegisterComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should be true'),()=>{
    expect(component.submitted).toBeTrue();
  }

  it('should be Registration From title '),()=>{
    expect(component.title).toEqual('Registration From');
  }

 
});
